+++ basic/drawing.c
@@ -1,5 +1,7 @@
 #include "basic.h"
 
+int line_width = 0;
+
 void colour(Uint8 r, Uint8 g, Uint8 b)
 {
 	SDL_SetRenderDrawColor(renderer, r, g, b, SDL_ALPHA_OPAQUE);
@@ -10,7 +12,10 @@ void line(int x0, int y0, int x1, int y1)
 {
 	Uint8 r, g, b, a;
 	SDL_GetRenderDrawColor(renderer, &r, &g, &b, &a);
-	lineRGBA(renderer, x0, y0, x1, y1, r, g, b, a);
+	if (line_width == 0)
+		lineRGBA(renderer, x0, y0, x1, y1, r, g, b, a);
+	else
+		thickLineRGBA(renderer, x0, y0, x1, y1, line_width, r, g, b, a);
 }
 
 void rect(int x, int y, int w, int h)